fix image vfs bug

+++ ports/k210-freertos/mpy_support/omv/img/apriltag.c
@@ -10398,7 +10398,7 @@ zarray_t *apriltag_quad_thresh(apriltag_detector_t *td, image_u8_t *im, bool ove
         do_unionfind_line(uf, threshim, h, w, ts, y);
     }
 
-    uint32_t nclustermap;
+    uint64_t nclustermap;
     struct uint32_zarray_entry **clustermap = fb_alloc0_all(&nclustermap);
     nclustermap /= sizeof(struct uint32_zarray_entry*);
     if (!nclustermap) fb_alloc_fail();
